# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CoinHub API - virtual coin economy platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/profile:
    patch:
      operationId: updateMyProfile
      tags: [user]
      summary: Profil maglumatlary täzelemek
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bio:
                  type: string
                  maxLength: 200
                avatarColor:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /me/transfer:
    post:
      operationId: transferCoins
      tags: [user]
      summary: Başga ulanyja TMT geçirmek
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipientPublicId:
                  type: string
                amount:
                  type: integer
                  minimum: 1
                note:
                  type: string
                  maxLength: 100
              required: [recipientPublicId, amount]
      responses:
        "200":
          description: Transfer success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"
...
  /admin/news:
    post:
      operationId: adminCreateNews
      tags: [admin]
      summary: Tazelik döretmek
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                body:
                  type: string
                  minLength: 1
                  maxLength: 1000
              required: [title, body]
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsPost"
  /admin/news/{newsId}:
    delete:
      operationId: adminDeleteNews
      tags: [admin]
      summary: Tazelik öçürmek
      parameters:
        - in: path
          name: newsId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    User:
      type: object
      properties:
        id:
          type: string
        publicId:
          type: string
        username:
          type: string
        coins:
          type: integer
        createdAt:
          type: string
        isAdmin:
          type: boolean
        email:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarColor:
          type: string
        bonusReady:
          type: boolean
        unreadNotifications:
          type: integer
        unreadDms:
          type: integer
      required: [id, publicId, username, coins, createdAt, isAdmin, avatarColor, bonusReady, unreadNotifications, unreadDms]
...
    PublicUser:
      type: object
      properties:
        id:
          type: string
        publicId:
          type: string
        username:
          type: string
        avatarColor:
          type: string
        isAdmin:
          type: boolean
      required: [id, publicId, username, avatarColor, isAdmin]
...
    TransferResult:
      type: object
      properties:
        ok:
          type: boolean
        senderBalance:
          type: integer
        recipient:
          $ref: "#/components/schemas/PublicUser"
        amount:
          type: integer
      required: [ok, senderBalance, recipient, amount]
...
    NewsPost:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
      required: [id, title, body, createdAt]
...