# lib/api-zod/src/generated/types/newsPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CoinHub API - virtual coin economy platform
 * OpenAPI spec version: 0.1.0
 */

export interface NewsPost {
  id: string;
  title: string;
  body: string;
  createdAt: string;
}